Allometric equations for linear models

Allometric equations that are considered in the development of linear models.

eqns_info

Format

A dataframe with 5 variables:

modeltype

One of the six general types of allometric equations.

base_equation

Basic form of the allometric equation.

base_formula

The basic formula used to fit the model.

weights

The weight argument in the model.

modelcode

Character string used to represent the unique combinations of equations and weights.

Details

Six allometric equations used to develop linear models for urban trees:

Linear: \(y_{i} = a + bx_{i} + \epsilon_{i}/\sqrt{w_{i}}\)

Quadratic: \(y_{i} = a + bx_{i} + cx^2_{i} + \epsilon_{i}/\sqrt{w_{i}}\)

Cubic: \(y_{i} = a + bx_{i} + cx^2_{i} + dx^3_{i} + \epsilon_{i}/\sqrt{w_{i}}\)

Quartic: \(y_{i} = a + bx_{i} + cx^2_{i} + dx^3_{i} + ex^4_{i} + \epsilon_{i}/\sqrt{w_{i}}\)

Log-log: \(\log{(y_{i})} = a + b(\log{(\log{(x_{i} + 1)})}) + \epsilon_{i}/\sqrt{w_{i}}\)

Exponential: \(\log{(y_{i})} = a + bx_{i} + 1 + \epsilon_{i}/\sqrt{w_{i}}\)

where

\(y_{i}\) = Response variable of individual tree \(i\), \(i = 1,2,3\)... \(n\), \(n\) = number of observations

\(x_{i}\) = Predictor variable

\(a,b,c,d,e\) = Parameters to be estimated

\(\epsilon_{i}\) = Normally distributed error term

\(w_{i}\) = Known weight that takes one of the four following forms: \(w_{i} = 1\), \(w_{i} = 1/\sqrt{x_{i}}\), \(w_{i} = 1/\sqrt{x_{i}}\), \(w_{i} = 1/x_{i}^2\).

References

McPherson E. G., van Doorn N. S. & Peper P. J. (2016) Urban Tree Database and Allometric Equations. General Technical Report PSW-GTR-253, USDA Forest Service, 86.

Examples

data(eqns_info)
head(eqns_info)

#>   modeltype    base_equation   base_formula      weights modelcode
#> 1    Linear       y = a + bx          y ~ x                 lin_w1
#> 2    Linear       y = a + bx          y ~ x I(1/sqrt(x))    lin_w2
#> 3    Linear       y = a + bx          y ~ x       I(1/x)    lin_w3
#> 4    Linear       y = a + bx          y ~ x     I(1/x^2)    lin_w4
#> 5 Quadratic y = a + bx + x^2 y ~ x + I(x^2)                quad_w1
#> 6 Quadratic y = a + bx + x^2 y ~ x + I(x^2) I(1/sqrt(x))   quad_w2