Helper function to generate new predictors for each species for simulation/prediction

Helper function to generate new predictors for each species for simulation/prediction

generate_newX(
  data,
  response = "height",
  predictor = "diameter",
  species = "species",
  length.out = 100
)

Arguments

data
Dataframe that contains the variables of interest. Each row is a measurement for an individual tree of a particular species.

response
Column name of the response variable in data. Defaults to height.

predictor
Column name of the predictor variable in data. Defaults to diameter.

species
Column name of the species variable in data. Defaults to species.

length.out
Number of new X values to generate for each species. Defaults to 100. Set a higher value for greater resolution at the cost of computational time.

Value

A new dataset containing the columns species, new X, and whether X is extrapolated for each species.

Examples

generate_newX(urbantrees,
              response = "height", predictor = "diameter",
              species = "species")

#>                           sp          x extrapolated
#> 101            Hopea odorata 0.02864789          Yes
#> 106            Hopea odorata 0.04379172           No
#> 111            Hopea odorata 0.05893556           No
#> 116            Hopea odorata 0.07407939           No
#> 121            Hopea odorata 0.08922323           No
#> 126            Hopea odorata 0.10436706           No
#> 131            Hopea odorata 0.11951089           No
#> 136            Hopea odorata 0.13465473           No
#> 141            Hopea odorata 0.14979856           No
#> 146            Hopea odorata 0.16494240           No
#> 151            Hopea odorata 0.18008623           No
#> 143            Hopea odorata 0.19523006           No
#> 148            Hopea odorata 0.21037390           No
#> 153            Hopea odorata 0.22551773           No
#> 158            Hopea odorata 0.24066157           No
#> 163            Hopea odorata 0.25580540           No
#> 168            Hopea odorata 0.27094923           No
#> 173            Hopea odorata 0.28609307           No
#> 178            Hopea odorata 0.30123690          Yes
#> 183            Hopea odorata 0.31638074          Yes
#> 188            Hopea odorata 0.33152457          Yes
#> 193            Hopea odorata 0.34666840          Yes
#> 198            Hopea odorata 0.36181224          Yes
#> 199            Hopea odorata 0.37695607          Yes
#> 195            Hopea odorata 0.39209991          Yes
#> 105            Hopea odorata 0.40724374          Yes
#> 110            Hopea odorata 0.42238757          Yes
#> 115            Hopea odorata 0.43753141          Yes
#> 120            Hopea odorata 0.45267524          Yes
#> 125            Hopea odorata 0.46781908          Yes
#> 117            Hopea odorata 0.48296291          Yes
#> 122            Hopea odorata 0.49810674          Yes
#> 127            Hopea odorata 0.51325058          Yes
#> 132            Hopea odorata 0.52839441          Yes
#> 137            Hopea odorata 0.54353825          Yes
#> 142            Hopea odorata 0.55868208          Yes
#> 147            Hopea odorata 0.57382591          Yes
#> 152            Hopea odorata 0.58896975          Yes
#> 157            Hopea odorata 0.60411358          Yes
#> 162            Hopea odorata 0.61925741          Yes
#> 167            Hopea odorata 0.63440125          Yes
#> 172            Hopea odorata 0.64954508          Yes
#> 177            Hopea odorata 0.66468892          Yes
#> 169            Hopea odorata 0.67983275          Yes
#> 174            Hopea odorata 0.69497658          Yes
#> 179            Hopea odorata 0.71012042          Yes
#> 184            Hopea odorata 0.72526425          Yes
#> 189            Hopea odorata 0.74040809          Yes
#> 194            Hopea odorata 0.75555192          Yes
#> 104            Hopea odorata 0.77069575          Yes
#> 109            Hopea odorata 0.78583959          Yes
#> 114            Hopea odorata 0.80098342          Yes
#> 119            Hopea odorata 0.81612726          Yes
#> 124            Hopea odorata 0.83127109          Yes
#> 129            Hopea odorata 0.84641492          Yes
#> 134            Hopea odorata 0.86155876          Yes
#> 139            Hopea odorata 0.87670259          Yes
#> 144            Hopea odorata 0.89184643          Yes
#> 149            Hopea odorata 0.90699026          Yes
#> 154            Hopea odorata 0.92213409          Yes
#> 159            Hopea odorata 0.93727793          Yes
#> 164            Hopea odorata 0.95242176          Yes
#> 156            Hopea odorata 0.96756560          Yes
#> 161            Hopea odorata 0.98270943          Yes
#> 166            Hopea odorata 0.99785326          Yes
#> 171            Hopea odorata 1.01299710          Yes
#> 176            Hopea odorata 1.02814093          Yes
#> 181            Hopea odorata 1.04328477          Yes
#> 186            Hopea odorata 1.05842860          Yes
#> 191            Hopea odorata 1.07357243          Yes
#> 196            Hopea odorata 1.08871627          Yes
#> 197            Hopea odorata 1.10386010          Yes
#> 200            Hopea odorata 1.11900394          Yes
#> 103            Hopea odorata 1.13414777          Yes
#> 108            Hopea odorata 1.14929160          Yes
#> 113            Hopea odorata 1.16443544          Yes
#> 118            Hopea odorata 1.17957927          Yes
#> 123            Hopea odorata 1.19472311          Yes
#> 128            Hopea odorata 1.20986694          Yes
#> 133            Hopea odorata 1.22501077          Yes
#> 138            Hopea odorata 1.24015461          Yes
#> 130            Hopea odorata 1.25529844          Yes
#> 135            Hopea odorata 1.27044228          Yes
#> 140            Hopea odorata 1.28558611          Yes
#> 145            Hopea odorata 1.30072994          Yes
#> 150            Hopea odorata 1.31587378          Yes
#> 155            Hopea odorata 1.33101761          Yes
#> 160            Hopea odorata 1.34616145          Yes
#> 165            Hopea odorata 1.36130528          Yes
#> 170            Hopea odorata 1.37644911          Yes
#> 175            Hopea odorata 1.39159295          Yes
#> 180            Hopea odorata 1.40673678          Yes
#> 185            Hopea odorata 1.42188062          Yes
#> 190            Hopea odorata 1.43702445          Yes
#> 182            Hopea odorata 1.45216828          Yes
#> 187            Hopea odorata 1.46731212          Yes
#> 192            Hopea odorata 1.48245595          Yes
#> 102            Hopea odorata 1.49759979          Yes
#> 107            Hopea odorata 1.51274362          Yes
#> 112            Hopea odorata 1.52788745          Yes
#> 423 Xanthostemon chrysanthus 0.02864789           No
#> 428 Xanthostemon chrysanthus 0.04379172           No
#> 433 Xanthostemon chrysanthus 0.05893556           No
#> 408 Xanthostemon chrysanthus 0.07407939           No
#> 401 Xanthostemon chrysanthus 0.08922323           No
#> 405 Xanthostemon chrysanthus 0.10436706           No
#> 402 Xanthostemon chrysanthus 0.11951089           No
#> 406 Xanthostemon chrysanthus 0.13465473           No
#> 407 Xanthostemon chrysanthus 0.14979856           No
#> 403 Xanthostemon chrysanthus 0.16494240           No
#> 404 Xanthostemon chrysanthus 0.18008623           No
#> 409 Xanthostemon chrysanthus 0.19523006           No
#> 414 Xanthostemon chrysanthus 0.21037390           No
#> 419 Xanthostemon chrysanthus 0.22551773           No
#> 411 Xanthostemon chrysanthus 0.24066157           No
#> 416 Xanthostemon chrysanthus 0.25580540           No
#> 421 Xanthostemon chrysanthus 0.27094923           No
#> 413 Xanthostemon chrysanthus 0.28609307           No
#> 418 Xanthostemon chrysanthus 0.30123690           No
#> 410 Xanthostemon chrysanthus 0.31638074           No
#> 415 Xanthostemon chrysanthus 0.33152457           No
#> 420 Xanthostemon chrysanthus 0.34666840           No
#> 412 Xanthostemon chrysanthus 0.36181224          Yes
#> 417 Xanthostemon chrysanthus 0.37695607          Yes
#> 422 Xanthostemon chrysanthus 0.39209991          Yes
#> 427 Xanthostemon chrysanthus 0.40724374          Yes
#> 432 Xanthostemon chrysanthus 0.42238757          Yes
#> 437 Xanthostemon chrysanthus 0.43753141          Yes
#> 442 Xanthostemon chrysanthus 0.45267524          Yes
#> 447 Xanthostemon chrysanthus 0.46781908          Yes
#> 439 Xanthostemon chrysanthus 0.48296291          Yes
#> 444 Xanthostemon chrysanthus 0.49810674          Yes
#> 449 Xanthostemon chrysanthus 0.51325058          Yes
#> 454 Xanthostemon chrysanthus 0.52839441          Yes
#> 459 Xanthostemon chrysanthus 0.54353825          Yes
#> 451 Xanthostemon chrysanthus 0.55868208          Yes
#> 456 Xanthostemon chrysanthus 0.57382591          Yes
#> 461 Xanthostemon chrysanthus 0.58896975          Yes
#> 466 Xanthostemon chrysanthus 0.60411358          Yes
#> 471 Xanthostemon chrysanthus 0.61925741          Yes
#> 476 Xanthostemon chrysanthus 0.63440125          Yes
#> 481 Xanthostemon chrysanthus 0.64954508          Yes
#> 486 Xanthostemon chrysanthus 0.66468892          Yes
#> 478 Xanthostemon chrysanthus 0.67983275          Yes
#> 483 Xanthostemon chrysanthus 0.69497658          Yes
#> 488 Xanthostemon chrysanthus 0.71012042          Yes
#> 493 Xanthostemon chrysanthus 0.72526425          Yes
#> 498 Xanthostemon chrysanthus 0.74040809          Yes
#> 490 Xanthostemon chrysanthus 0.75555192          Yes
#> 426 Xanthostemon chrysanthus 0.77069575          Yes
#> 431 Xanthostemon chrysanthus 0.78583959          Yes
#> 436 Xanthostemon chrysanthus 0.80098342          Yes
#> 441 Xanthostemon chrysanthus 0.81612726          Yes
#> 446 Xanthostemon chrysanthus 0.83127109          Yes
#> 438 Xanthostemon chrysanthus 0.84641492          Yes
#> 443 Xanthostemon chrysanthus 0.86155876          Yes
#> 448 Xanthostemon chrysanthus 0.87670259          Yes
#> 453 Xanthostemon chrysanthus 0.89184643          Yes
#> 458 Xanthostemon chrysanthus 0.90699026          Yes
#> 463 Xanthostemon chrysanthus 0.92213409          Yes
#> 468 Xanthostemon chrysanthus 0.93727793          Yes
#> 473 Xanthostemon chrysanthus 0.95242176          Yes
#> 465 Xanthostemon chrysanthus 0.96756560          Yes
#> 470 Xanthostemon chrysanthus 0.98270943          Yes
#> 475 Xanthostemon chrysanthus 0.99785326          Yes
#> 480 Xanthostemon chrysanthus 1.01299710          Yes
#> 485 Xanthostemon chrysanthus 1.02814093          Yes
#> 477 Xanthostemon chrysanthus 1.04328477          Yes
#> 482 Xanthostemon chrysanthus 1.05842860          Yes
#> 487 Xanthostemon chrysanthus 1.07357243          Yes
#> 492 Xanthostemon chrysanthus 1.08871627          Yes
#> 497 Xanthostemon chrysanthus 1.10386010          Yes
#> 500 Xanthostemon chrysanthus 1.11900394          Yes
#> 425 Xanthostemon chrysanthus 1.13414777          Yes
#> 430 Xanthostemon chrysanthus 1.14929160          Yes
#> 435 Xanthostemon chrysanthus 1.16443544          Yes
#> 440 Xanthostemon chrysanthus 1.17957927          Yes
#> 445 Xanthostemon chrysanthus 1.19472311          Yes
#> 450 Xanthostemon chrysanthus 1.20986694          Yes
#> 455 Xanthostemon chrysanthus 1.22501077          Yes
#> 460 Xanthostemon chrysanthus 1.24015461          Yes
#> 452 Xanthostemon chrysanthus 1.25529844          Yes
#> 457 Xanthostemon chrysanthus 1.27044228          Yes
#> 462 Xanthostemon chrysanthus 1.28558611          Yes
#> 467 Xanthostemon chrysanthus 1.30072994          Yes
#> 472 Xanthostemon chrysanthus 1.31587378          Yes
#> 464 Xanthostemon chrysanthus 1.33101761          Yes
#> 469 Xanthostemon chrysanthus 1.34616145          Yes
#> 474 Xanthostemon chrysanthus 1.36130528          Yes
#> 479 Xanthostemon chrysanthus 1.37644911          Yes
#> 484 Xanthostemon chrysanthus 1.39159295          Yes
#> 489 Xanthostemon chrysanthus 1.40673678          Yes
#> 494 Xanthostemon chrysanthus 1.42188062          Yes
#> 495 Xanthostemon chrysanthus 1.43702445          Yes
#> 491 Xanthostemon chrysanthus 1.45216828          Yes
#> 496 Xanthostemon chrysanthus 1.46731212          Yes
#> 499 Xanthostemon chrysanthus 1.48245595          Yes
#> 424 Xanthostemon chrysanthus 1.49759979          Yes
#> 429 Xanthostemon chrysanthus 1.51274362          Yes
#> 434 Xanthostemon chrysanthus 1.52788745          Yes
#> 345       Terminalia mantaly 0.02864789          Yes
#> 350       Terminalia mantaly 0.04379172           No
#> 355       Terminalia mantaly 0.05893556           No
#> 304       Terminalia mantaly 0.07407939           No
#> 309       Terminalia mantaly 0.08922323           No
#> 314       Terminalia mantaly 0.10436706           No
#> 306       Terminalia mantaly 0.11951089           No
#> 311       Terminalia mantaly 0.13465473           No
#> 316       Terminalia mantaly 0.14979856           No
#> 308       Terminalia mantaly 0.16494240           No
#> 313       Terminalia mantaly 0.18008623           No
#> 318       Terminalia mantaly 0.19523006           No
#> 323       Terminalia mantaly 0.21037390           No
#> 328       Terminalia mantaly 0.22551773           No
#> 320       Terminalia mantaly 0.24066157           No
#> 325       Terminalia mantaly 0.25580540           No
#> 330       Terminalia mantaly 0.27094923           No
#> 335       Terminalia mantaly 0.28609307           No
#> 340       Terminalia mantaly 0.30123690           No
#> 332       Terminalia mantaly 0.31638074           No
#> 337       Terminalia mantaly 0.33152457           No
#> 342       Terminalia mantaly 0.34666840           No
#> 334       Terminalia mantaly 0.36181224           No
#> 339       Terminalia mantaly 0.37695607           No
#> 344       Terminalia mantaly 0.39209991           No
#> 349       Terminalia mantaly 0.40724374           No
#> 354       Terminalia mantaly 0.42238757           No
#> 303       Terminalia mantaly 0.43753141           No
#> 301       Terminalia mantaly 0.45267524           No
#> 302       Terminalia mantaly 0.46781908           No
#> 305       Terminalia mantaly 0.48296291           No
#> 310       Terminalia mantaly 0.49810674           No
#> 315       Terminalia mantaly 0.51325058           No
#> 307       Terminalia mantaly 0.52839441           No
#> 312       Terminalia mantaly 0.54353825           No
#> 317       Terminalia mantaly 0.55868208           No
#> 322       Terminalia mantaly 0.57382591          Yes
#> 327       Terminalia mantaly 0.58896975          Yes
#> 319       Terminalia mantaly 0.60411358          Yes
#> 324       Terminalia mantaly 0.61925741          Yes
#> 329       Terminalia mantaly 0.63440125          Yes
#> 321       Terminalia mantaly 0.64954508          Yes
#> 326       Terminalia mantaly 0.66468892          Yes
#> 331       Terminalia mantaly 0.67983275          Yes
#> 336       Terminalia mantaly 0.69497658          Yes
#> 341       Terminalia mantaly 0.71012042          Yes
#> 333       Terminalia mantaly 0.72526425          Yes
#> 338       Terminalia mantaly 0.74040809          Yes
#> 343       Terminalia mantaly 0.75555192          Yes
#> 348       Terminalia mantaly 0.77069575          Yes
#> 353       Terminalia mantaly 0.78583959          Yes
#> 358       Terminalia mantaly 0.80098342          Yes
#> 363       Terminalia mantaly 0.81612726          Yes
#> 368       Terminalia mantaly 0.83127109          Yes
#> 360       Terminalia mantaly 0.84641492          Yes
#> 365       Terminalia mantaly 0.86155876          Yes
#> 370       Terminalia mantaly 0.87670259          Yes
#> 375       Terminalia mantaly 0.89184643          Yes
#> 380       Terminalia mantaly 0.90699026          Yes
#> 372       Terminalia mantaly 0.92213409          Yes
#> 377       Terminalia mantaly 0.93727793          Yes
#> 382       Terminalia mantaly 0.95242176          Yes
#> 374       Terminalia mantaly 0.96756560          Yes
#> 379       Terminalia mantaly 0.98270943          Yes
#> 384       Terminalia mantaly 0.99785326          Yes
#> 389       Terminalia mantaly 1.01299710          Yes
#> 394       Terminalia mantaly 1.02814093          Yes
#> 386       Terminalia mantaly 1.04328477          Yes
#> 391       Terminalia mantaly 1.05842860          Yes
#> 396       Terminalia mantaly 1.07357243          Yes
#> 399       Terminalia mantaly 1.08871627          Yes
#> 400       Terminalia mantaly 1.10386010          Yes
#> 398       Terminalia mantaly 1.11900394          Yes
#> 347       Terminalia mantaly 1.13414777          Yes
#> 352       Terminalia mantaly 1.14929160          Yes
#> 357       Terminalia mantaly 1.16443544          Yes
#> 362       Terminalia mantaly 1.17957927          Yes
#> 367       Terminalia mantaly 1.19472311          Yes
#> 359       Terminalia mantaly 1.20986694          Yes
#> 364       Terminalia mantaly 1.22501077          Yes
#> 369       Terminalia mantaly 1.24015461          Yes
#> 361       Terminalia mantaly 1.25529844          Yes
#> 366       Terminalia mantaly 1.27044228          Yes
#> 371       Terminalia mantaly 1.28558611          Yes
#> 376       Terminalia mantaly 1.30072994          Yes
#> 381       Terminalia mantaly 1.31587378          Yes
#> 373       Terminalia mantaly 1.33101761          Yes
#> 378       Terminalia mantaly 1.34616145          Yes
#> 383       Terminalia mantaly 1.36130528          Yes
#> 388       Terminalia mantaly 1.37644911          Yes
#> 393       Terminalia mantaly 1.39159295          Yes
#> 385       Terminalia mantaly 1.40673678          Yes
#> 390       Terminalia mantaly 1.42188062          Yes
#> 395       Terminalia mantaly 1.43702445          Yes
#> 387       Terminalia mantaly 1.45216828          Yes
#> 392       Terminalia mantaly 1.46731212          Yes
#> 397       Terminalia mantaly 1.48245595          Yes
#> 346       Terminalia mantaly 1.49759979          Yes
#> 351       Terminalia mantaly 1.51274362          Yes
#> 356       Terminalia mantaly 1.52788745          Yes
#> 267     Syzygium myrtifolium 0.02864789          Yes
#> 272     Syzygium myrtifolium 0.04379172           No
#> 277     Syzygium myrtifolium 0.05893556           No
#> 202     Syzygium myrtifolium 0.07407939           No
#> 205     Syzygium myrtifolium 0.08922323           No
#> 210     Syzygium myrtifolium 0.10436706           No
#> 206     Syzygium myrtifolium 0.11951089           No
#> 207     Syzygium myrtifolium 0.13465473           No
#> 212     Syzygium myrtifolium 0.14979856           No
#> 217     Syzygium myrtifolium 0.16494240           No
#> 222     Syzygium myrtifolium 0.18008623           No
#> 227     Syzygium myrtifolium 0.19523006           No
#> 232     Syzygium myrtifolium 0.21037390           No
#> 237     Syzygium myrtifolium 0.22551773           No
#> 229     Syzygium myrtifolium 0.24066157           No
#> 234     Syzygium myrtifolium 0.25580540           No
#> 239     Syzygium myrtifolium 0.27094923           No
#> 244     Syzygium myrtifolium 0.28609307           No
#> 249     Syzygium myrtifolium 0.30123690           No
#> 241     Syzygium myrtifolium 0.31638074           No
#> 246     Syzygium myrtifolium 0.33152457           No
#> 251     Syzygium myrtifolium 0.34666840           No
#> 256     Syzygium myrtifolium 0.36181224           No
#> 261     Syzygium myrtifolium 0.37695607           No
#> 266     Syzygium myrtifolium 0.39209991           No
#> 271     Syzygium myrtifolium 0.40724374           No
#> 276     Syzygium myrtifolium 0.42238757           No
#> 201     Syzygium myrtifolium 0.43753141           No
#> 204     Syzygium myrtifolium 0.45267524           No
#> 209     Syzygium myrtifolium 0.46781908           No
#> 214     Syzygium myrtifolium 0.48296291           No
#> 219     Syzygium myrtifolium 0.49810674           No
#> 224     Syzygium myrtifolium 0.51325058           No
#> 216     Syzygium myrtifolium 0.52839441           No
#> 221     Syzygium myrtifolium 0.54353825           No
#> 226     Syzygium myrtifolium 0.55868208           No
#> 231     Syzygium myrtifolium 0.57382591          Yes
#> 236     Syzygium myrtifolium 0.58896975          Yes
#> 228     Syzygium myrtifolium 0.60411358          Yes
#> 233     Syzygium myrtifolium 0.61925741          Yes
#> 238     Syzygium myrtifolium 0.63440125          Yes
#> 243     Syzygium myrtifolium 0.64954508          Yes
#> 248     Syzygium myrtifolium 0.66468892          Yes
#> 253     Syzygium myrtifolium 0.67983275          Yes
#> 258     Syzygium myrtifolium 0.69497658          Yes
#> 263     Syzygium myrtifolium 0.71012042          Yes
#> 255     Syzygium myrtifolium 0.72526425          Yes
#> 260     Syzygium myrtifolium 0.74040809          Yes
#> 265     Syzygium myrtifolium 0.75555192          Yes
#> 270     Syzygium myrtifolium 0.77069575          Yes
#> 275     Syzygium myrtifolium 0.78583959          Yes
#> 211     Syzygium myrtifolium 0.80098342          Yes
#> 203     Syzygium myrtifolium 0.81612726          Yes
#> 208     Syzygium myrtifolium 0.83127109          Yes
#> 213     Syzygium myrtifolium 0.84641492          Yes
#> 218     Syzygium myrtifolium 0.86155876          Yes
#> 223     Syzygium myrtifolium 0.87670259          Yes
#> 215     Syzygium myrtifolium 0.89184643          Yes
#> 220     Syzygium myrtifolium 0.90699026          Yes
#> 225     Syzygium myrtifolium 0.92213409          Yes
#> 230     Syzygium myrtifolium 0.93727793          Yes
#> 235     Syzygium myrtifolium 0.95242176          Yes
#> 240     Syzygium myrtifolium 0.96756560          Yes
#> 245     Syzygium myrtifolium 0.98270943          Yes
#> 250     Syzygium myrtifolium 0.99785326          Yes
#> 242     Syzygium myrtifolium 1.01299710          Yes
#> 247     Syzygium myrtifolium 1.02814093          Yes
#> 252     Syzygium myrtifolium 1.04328477          Yes
#> 257     Syzygium myrtifolium 1.05842860          Yes
#> 262     Syzygium myrtifolium 1.07357243          Yes
#> 254     Syzygium myrtifolium 1.08871627          Yes
#> 259     Syzygium myrtifolium 1.10386010          Yes
#> 264     Syzygium myrtifolium 1.11900394          Yes
#> 269     Syzygium myrtifolium 1.13414777          Yes
#> 274     Syzygium myrtifolium 1.14929160          Yes
#> 279     Syzygium myrtifolium 1.16443544          Yes
#> 284     Syzygium myrtifolium 1.17957927          Yes
#> 289     Syzygium myrtifolium 1.19472311          Yes
#> 281     Syzygium myrtifolium 1.20986694          Yes
#> 286     Syzygium myrtifolium 1.22501077          Yes
#> 291     Syzygium myrtifolium 1.24015461          Yes
#> 283     Syzygium myrtifolium 1.25529844          Yes
#> 288     Syzygium myrtifolium 1.27044228          Yes
#> 280     Syzygium myrtifolium 1.28558611          Yes
#> 285     Syzygium myrtifolium 1.30072994          Yes
#> 290     Syzygium myrtifolium 1.31587378          Yes
#> 282     Syzygium myrtifolium 1.33101761          Yes
#> 287     Syzygium myrtifolium 1.34616145          Yes
#> 292     Syzygium myrtifolium 1.36130528          Yes
#> 297     Syzygium myrtifolium 1.37644911          Yes
#> 298     Syzygium myrtifolium 1.39159295          Yes
#> 294     Syzygium myrtifolium 1.40673678          Yes
#> 295     Syzygium myrtifolium 1.42188062          Yes
#> 299     Syzygium myrtifolium 1.43702445          Yes
#> 296     Syzygium myrtifolium 1.45216828          Yes
#> 300     Syzygium myrtifolium 1.46731212          Yes
#> 293     Syzygium myrtifolium 1.48245595          Yes
#> 268     Syzygium myrtifolium 1.49759979          Yes
#> 273     Syzygium myrtifolium 1.51274362          Yes
#> 278     Syzygium myrtifolium 1.52788745          Yes
#> 89             Albizia saman 0.02864789          Yes
#> 94             Albizia saman 0.04379172          Yes
#> 99             Albizia saman 0.05893556          Yes
#> 9              Albizia saman 0.07407939          Yes
#> 14             Albizia saman 0.08922323          Yes
#> 19             Albizia saman 0.10436706          Yes
#> 11             Albizia saman 0.11951089          Yes
#> 16             Albizia saman 0.13465473          Yes
#> 21             Albizia saman 0.14979856          Yes
#> 26             Albizia saman 0.16494240          Yes
#> 31             Albizia saman 0.18008623          Yes
#> 36             Albizia saman 0.19523006          Yes
#> 41             Albizia saman 0.21037390          Yes
#> 46             Albizia saman 0.22551773          Yes
#> 51             Albizia saman 0.24066157          Yes
#> 56             Albizia saman 0.25580540          Yes
#> 61             Albizia saman 0.27094923          Yes
#> 66             Albizia saman 0.28609307          Yes
#> 71             Albizia saman 0.30123690          Yes
#> 63             Albizia saman 0.31638074           No
#> 68             Albizia saman 0.33152457           No
#> 73             Albizia saman 0.34666840           No
#> 78             Albizia saman 0.36181224           No
#> 83             Albizia saman 0.37695607           No
#> 88             Albizia saman 0.39209991           No
#> 93             Albizia saman 0.40724374           No
#> 98             Albizia saman 0.42238757           No
#> 1              Albizia saman 0.43753141           No
#> 4              Albizia saman 0.45267524           No
#> 5              Albizia saman 0.46781908           No
#> 10             Albizia saman 0.48296291           No
#> 15             Albizia saman 0.49810674           No
#> 20             Albizia saman 0.51325058           No
#> 25             Albizia saman 0.52839441           No
#> 30             Albizia saman 0.54353825           No
#> 35             Albizia saman 0.55868208           No
#> 40             Albizia saman 0.57382591           No
#> 45             Albizia saman 0.58896975           No
#> 37             Albizia saman 0.60411358           No
#> 42             Albizia saman 0.61925741           No
#> 47             Albizia saman 0.63440125           No
#> 52             Albizia saman 0.64954508           No
#> 57             Albizia saman 0.66468892           No
#> 62             Albizia saman 0.67983275           No
#> 67             Albizia saman 0.69497658           No
#> 72             Albizia saman 0.71012042           No
#> 77             Albizia saman 0.72526425           No
#> 82             Albizia saman 0.74040809           No
#> 87             Albizia saman 0.75555192           No
#> 92             Albizia saman 0.77069575           No
#> 97             Albizia saman 0.78583959           No
#> 7              Albizia saman 0.80098342           No
#> 12             Albizia saman 0.81612726           No
#> 17             Albizia saman 0.83127109           No
#> 22             Albizia saman 0.84641492           No
#> 27             Albizia saman 0.86155876           No
#> 32             Albizia saman 0.87670259           No
#> 24             Albizia saman 0.89184643           No
#> 29             Albizia saman 0.90699026           No
#> 34             Albizia saman 0.92213409           No
#> 39             Albizia saman 0.93727793           No
#> 44             Albizia saman 0.95242176           No
#> 49             Albizia saman 0.96756560           No
#> 54             Albizia saman 0.98270943           No
#> 59             Albizia saman 0.99785326           No
#> 64             Albizia saman 1.01299710           No
#> 69             Albizia saman 1.02814093           No
#> 74             Albizia saman 1.04328477           No
#> 79             Albizia saman 1.05842860           No
#> 84             Albizia saman 1.07357243           No
#> 76             Albizia saman 1.08871627           No
#> 81             Albizia saman 1.10386010           No
#> 86             Albizia saman 1.11900394           No
#> 91             Albizia saman 1.13414777           No
#> 96             Albizia saman 1.14929160           No
#> 6              Albizia saman 1.16443544           No
#> 2              Albizia saman 1.17957927           No
#> 3              Albizia saman 1.19472311           No
#> 8              Albizia saman 1.20986694           No
#> 13             Albizia saman 1.22501077           No
#> 18             Albizia saman 1.24015461           No
#> 23             Albizia saman 1.25529844           No
#> 28             Albizia saman 1.27044228           No
#> 33             Albizia saman 1.28558611           No
#> 38             Albizia saman 1.30072994           No
#> 43             Albizia saman 1.31587378           No
#> 48             Albizia saman 1.33101761           No
#> 53             Albizia saman 1.34616145           No
#> 58             Albizia saman 1.36130528           No
#> 50             Albizia saman 1.37644911           No
#> 55             Albizia saman 1.39159295           No
#> 60             Albizia saman 1.40673678           No
#> 65             Albizia saman 1.42188062           No
#> 70             Albizia saman 1.43702445           No
#> 75             Albizia saman 1.45216828           No
#> 80             Albizia saman 1.46731212           No
#> 85             Albizia saman 1.48245595           No
#> 90             Albizia saman 1.49759979           No
#> 95             Albizia saman 1.51274362           No
#> 100            Albizia saman 1.52788745           No