Fit linear mixed-effects (multilevel) model for all species

Fit linear mixed-effects (multilevel) model with 'species' specified as the random effect, using the lme4::lmer function under the hood. The full list of allometric equations that are considered can be found in ?eqns_info and data(eqns_info).

me_modelfit(
  data,
  species = "species",
  response = "height",
  predictor = "diameter"
)

Arguments

data
Dataframe that contains the variables of interest. Each row is a measurement for an individual tree.

species
Column name of the species variable in data. Defaults to species.

response
Column name of the response variable. Defaults to height.

predictor
Column name of the predictor variable. Defaults to diameter.

Value

A list of 5 elements:

models_rank

A model selection table of all types of mixed-effects models

best_model

The best model object.

R2

The conditional and marginal pseudo-\(R^2\) of the best model.

CF

Correction factor to adjust prediction or simulation.

warnings

Warning messages, if any, spit from the models. These usually indicate failure of model convergence.

Examples

data(urbantrees)
urbantrees <- subset(urbantrees, height < 40) # remove outliers

results <- me_modelfit(data = urbantrees,
                       response = "height", predictor = "diameter",
                       species = "species")

#> Fitting lin model with weight w1

#> Fitting lin model with weight w2

#> Fitting lin model with weight w3

#> Fitting lin model with weight w4

#> Fitting quad model with weight w1

#> Fitting quad model with weight w2

#> boundary (singular) fit: see ?isSingular

#> Fitting quad model with weight w3

#> boundary (singular) fit: see ?isSingular

#> Fitting quad model with weight w4

#> boundary (singular) fit: see ?isSingular

#> Fitting cub model with weight w1

#> boundary (singular) fit: see ?isSingular

#> Fitting cub model with weight w2

#> boundary (singular) fit: see ?isSingular

#> Fitting cub model with weight w3

#> boundary (singular) fit: see ?isSingular

#> Fitting cub model with weight w4

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> boundary (singular) fit: see ?isSingular

#> Fitting quart model with weight w1

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> boundary (singular) fit: see ?isSingular

#> Fitting quart model with weight w2

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> boundary (singular) fit: see ?isSingular

#> Fitting quart model with weight w3

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> boundary (singular) fit: see ?isSingular

#> Fitting quart model with weight w4

#> Warning: convergence code 1 from bobyqa: bobyqa -- maximum number of function evaluations exceeded

#> boundary (singular) fit: see ?isSingular

#> Fitting loglog model with weight w1

#> boundary (singular) fit: see ?isSingular

#> Fitting loglog model with weight w2

#> Fitting loglog model with weight w3

#> Fitting loglog model with weight w4

#> Fitting expo model with weight w1

#> Fitting expo model with weight w2

#> Fitting expo model with weight w3

#> Fitting expo model with weight w4

#> Opps, some models have warning messages. See 'warnings' in the output list.

#> Warning: response differs between models

#> Warning: 'r.squaredGLMM' now calculates a revised statistic. See the help page.

results

#> $models_rank
#>           df    logLik     AICc       delta
#> loglog_w2  6 -2958.738 5929.529    0.000000
#> loglog_w1  6 -2960.885 5933.823    4.294174
#> loglog_w3  6 -3034.235 6080.524  150.994822
#> quad_w1   10 -3043.290 6106.719  177.190202
#> cub_w1    15 -3045.773 6121.852  192.322669
#> quart_w1  21 -3041.152 6124.895  195.365797
#> expo_w2    6 -3058.214 6128.482  198.952760
#> expo_w1    6 -3094.605 6201.264  271.734624
#> lin_w1     6 -3102.162 6216.377  286.848075
#> expo_w3    6 -3106.454 6224.960  295.431378
#> quad_w2   10 -3216.611 6453.362  523.833016
#> cub_w2    15 -3213.299 6456.904  527.374953
#> quart_w2  21 -3209.829 6462.250  532.720594
#> lin_w2     6 -3258.497 6529.047  599.518037
#> loglog_w4  6 -3467.702 6947.458 1017.928816
#> quad_w3   10 -3475.754 6971.648 1042.118758
#> cub_w3    15 -3476.526 6983.358 1053.829353
#> quart_w3  21 -3473.939 6990.469 1060.939858
#> lin_w3     6 -3513.926 7039.905 1110.375817
#> expo_w4    6 -3520.118 7052.290 1122.760909
#> quad_w4   10 -4283.323 8586.787 2657.257613
#> cub_w4    15 -4280.828 8591.963 2662.433904
#> quart_w4  21 -4279.595 8601.782 2672.252997
#> lin_w4     6 -4317.799 8647.652 2718.123088
#> 
#> $best_model
#> Linear mixed model fit by REML ['lmerMod']
#> Formula: y_trans ~ 1 + I(log(log(x + 1))) + (1 + I(log(log(x + 1))) |      sp)
#> Weights: weight.list[[best_mod_weight]]
#> REML criterion at convergence: 12.2458
#> Random effects:
#>  Groups   Name               Std.Dev. Corr
#>  sp       (Intercept)        0.17460      
#>           I(log(log(x + 1))) 0.06829  0.99
#>  Residual                    0.15308      
#> Number of obs: 1584, groups:  sp, 5
#> Fixed Effects:
#>        (Intercept)  I(log(log(x + 1)))  
#>             3.0575              0.6214  
#> 
#> $R2
#>            R2m       R2c
#> [1,] 0.8481846 0.8726846
#> 
#> $CF
#> [1] 1.0296
#> 
#> $warnings
#> $warnings$quad_w2
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quad_w3
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quad_w4
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$cub_w1
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$cub_w2
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$cub_w3
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$cub_w4
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quart_w1
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quart_w2
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quart_w3
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$quart_w4
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#> $warnings$loglog_w1
#> [1] "boundary (singular) fit: see ?isSingular"
#> 
#>