Fit data to pre-selected models across multiple species

Wrapper to run sp_modelfit() across multiple species. Data is fit to pre-selected allometric equations for each species, as defined within ref_table. The full list of allometric equations that may be considered in ref_table can be found in ?eqns_info and data(eqns_info).

sp_modelfit_multi(
  data,
  ref_table,
  species = "species",
  modelcode = "modelcode",
  response = "height",
  predictor = "diameter"
)

Arguments

data
Dataframe that contains the variables of interest. Each row is a measurement for an individual tree of a particular species.

ref_table
Dataframe containing an allometric equation for each tree species, in the form of a model code. Each row is a unique species.

species
Column name of the species variable in both the dataframes data and ref_table. Defaults to species.

modelcode
Column name containing the model codes in ref_table. Refer to data(eqns_info) for more information on model codes.

response
Column name of the response variable in data. Defaults to height.

predictor
Column name of the predictor variable in data. Defaults to diameter.

Value

A list of 2 elements:

List of each species' resulting model object, named sp_models.

Table showing each species' resulting model information, named sp_models_info.

sp_models_info

More information on output columns:

species: Name of tree species

modelcode: Model code for the best-fit equation

a, b, c, d, e: Parameter estimates

response_geom_mean: Geometric mean of the response variable used in calculation of AICc (only for transformed models)

correctn_factor: Bias correction factor to use on model predictions (only for transformed models)

predictor_min, predictor_max: Range of the predictor variable within the data used to generate the model (i.e. threshold values for model extrapolation)

residual_SE: Residual standard error of the model

mean_SE: Mean standard error of the model

adj_R2: Adjusted \(R^2\) of the model

n: Sample size (no. of trees used to fit model)

See also

Examples

# first select best-fit model for all species in data
data(urbantrees)
selected <- sp_modelselect_multi(urbantrees, species = 'species', response = 'height', predictor = 'diameter')

# use function
results <- sp_modelfit_multi(
  urbantrees, # any dataset with similar species can be used (in this case, we re-use the same data)
  ref_table = selected$sp_models_info, # refer to reference table from results above
  species = 'species', modelcode = 'modelcode',
  response = 'height', predictor = 'diameter'
)

results$sp_models[[1]] # model object for first species in list

#> 
#> Call:
#> lm(formula = data[[response]] ~ data[[predictor]] + I(data[[predictor]]^2))
#> 
#> Coefficients:
#>            (Intercept)       data[[predictor]]  I(data[[predictor]]^2)  
#>                  2.741                  19.664                  -6.061  
#> 

results$sp_models_info # summary of fitted models

#>                    species modelcode         a           b            c
#> 1            Albizia saman   quad_w1  2.740553  19.6642893    -6.061372
#> 2            Hopea odorata  quart_w4 -4.574237 279.2780006 -3105.195599
#> 3     Syzygium myrtifolium  quart_w2 -9.743595 275.2010554 -1580.335018
#> 4       Terminalia mantaly    cub_w3  3.730866   0.1733303   129.908835
#> 5 Xanthostemon chrysanthus   quad_w1  2.033997  32.4645629   -36.346261
#>            d          e response_geom_mean correctn_factor predictor_min
#> 1         NA         NA          13.454710               1    0.31194369
#> 2 15516.9602 -26692.934           5.354806               1    0.03183099
#> 3  3917.9761  -3398.366           8.037675               1    0.07321127
#> 4  -155.1424         NA           7.609798               1    0.03501409
#> 5         NA         NA           5.106066               1    0.04456338
#>   predictor_max residual_SE mean_SE adj_R2   n
#> 1     1.5278875      2.2412  4.8721 0.4499 100
#> 2     0.2864789     12.5967  2.9191 0.6010 100
#> 3     0.5188451      2.2546  2.4289 0.4341 100
#> 4     0.5602254      3.9638  3.1954 0.7588 100
#> 5     0.3501409      1.1312  1.2413 0.5887 100