Select a best-fit model for one species

Select a best-fit equation for one species, based on the lowest bias-corrected Aikaike’s information criterion (AICc).

sp_modelselect(data, response = "height", predictor = "diameter")

Arguments

data
Dataframe that contains the variables of interest. Each row is a measurement for an individual tree.

response
Column name of the response variable. Defaults to height.

predictor
Column name of the predictor variable. Defaults to diameter.

Value

A list of 3 elements:

Table showing models ranked by AICc value, named all_models_rank.

Best-fit model object, named best_model.

Table showing information on the best-fit model, named best_model_info.

best_model_info

More information on output columns:

modelcode: Model code for the best-fit equation

a, b, c, d, e: Parameter estimates

response_geom_mean: Geometric mean of the response variable used in calculation of AICc (only for transformed models)

correctn_factor: Bias correction factor to use on model predictions (only for transformed models)

predictor_min, predictor_max: Range of the predictor variable within the data used to generate the model (i.e. threshold values for model extrapolation)

residual_SE: Residual standard error of the model

mean_SE: Mean standard error of the model

adj_R2: Adjusted \(R^2\) of the model

n: Sample size (no. of trees used to fit model)

Details

All allometric equations considered (and ranked) can be found in ?eqns_info and data(eqns_info). To make the AICc values of equations with a transformed response variable comparable to untransformed equations, \(log(y_{i})\) is multiplied by the geometric mean of the response variable in data.

References

McPherson E. G., van Doorn N. S. & Peper P. J. (2016) Urban Tree Database and Allometric Equations. General Technical Report PSW-GTR-253, USDA Forest Service, 86.

Xiao, X., White, E. P., Hooten, M. B., & Durham, S. L. (2011). On the use of log-transformation vs. nonlinear regression for analyzing biological power laws. Ecology, 92(10), 1887–1894.

Burnham, K. P., & Anderson, D. R. (2004). Multimodel inference: Understanding AIC and BIC in model selection. Sociological Methods and Research, 33(2), 261–304.

See also

Examples

data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]  # subset data for one species
results <- sp_modelselect(Alb_sam, response = 'height', predictor = 'diameter')

head(results$all_models_rank)

#>   df     AICc    model
#> 1  3 591.9422   lin_w1
#> 2  4 593.0074  quad_w1
#> 3  5 593.4045   cub_w1
#> 4  6 594.0030 quart_w1
#> 5  3 594.0139   lin_w2
#> 6  5 595.5983   cub_w2

results$best_model

#> 
#> Call:
#> lm(formula = data[[response]] ~ data[[predictor]])
#> 
#> Coefficients:
#>       (Intercept)  data[[predictor]]  
#>             6.717              9.464  
#> 

results$best_model_info

#>   modelcode        a        b  c  d  e response_geom_mean correctn_factor
#> 1    lin_w1 6.717431 9.464096 NA NA NA           13.56608               1
#>   predictor_min predictor_max residual_SE mean_SE adj_R2   n
#> 1     0.3119437      1.527887       2.205  4.7889 0.4276 133