Heights and diameter sizes of 1,585 street trees in Singapore

A sample dataset containing the height and trunk diameter of five tree species, planted along streets in Singapore. Each row is a measurement for an individual tree.

urbantrees

Format

A dataframe with 1585 rows and 3 variables:

species

Name of tree species.

height

Tree height based on visual estimation by an arborist (m).

diameter

Tree girth measured with diameter tape (at breast-height), and converted to diameter size (m).

Source

Independently sampled validation data reported in:

Song, X. P., Lai, H. R., Wijedasa, L. S., Tan, P. Y., Edwards, P. J., & Richards, D. R. (2020), Height–diameter allometry for the management of city trees in the tropics. Environmental Research Letters, 15, 114017.

Examples

data(urbantrees)
unique(urbantrees$species) # species in the dataset
#> [1] "Hopea odorata"            "Xanthostemon chrysanthus"
#> [3] "Terminalia mantaly"       "Syzygium myrtifolium"    
#> [5] "Albizia saman"           

table(urbantrees$species) # number of data points (rows) per species
#> 
#>            Albizia saman            Hopea odorata     Syzygium myrtifolium 
#>                      133                      483                      353 
#>       Terminalia mantaly Xanthostemon chrysanthus 
#>                      197                      418